MVP achieved - btn click pastes rich text! :tada:
With the use of the clipboard API, we can copy the rich text from the
quill editor to the clipboard, then when the Medium Signature button
is pressed in an editing story, the text will be pasted and formatted
properly.

`clipboard-polyfill` was added to write rich text to the editor.