Implemented handling for non-existent signature
Previously, when no signature existed in storage and the user attempted
to use the nib icon button on Medium, nothing would happen. Thus, the
user wouldn't know why no signature is being pasted in their article.

This commit makes it so that if the above situation occurs, then the
settings page will open with a warning telling the user to add a signature
before continuing.