Implement / refactor notifications in the settings page
Fixes #6. Previously, there was a single notification that would warn the user to
add a signature before trying to use the signature button in Medium.

This adds a new notification (`successNotification`) that appears whenever
the user saves their signature, adding to the overall UX. With the addition
of this new notification, some minor refactoring was done.